% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOmap.R
\name{buildGOmap}
\alias{buildGOmap}
\title{buildGOmap}
\usage{
buildGOmap(TERM2GENE)
}
\arguments{
\item{TERM2GENE}{data.frame with two or three columns of GO TERM, GENE and ONTOLOGY (optional)}
}
\value{
data.frame, GO annotation with direct and indirect annotation
}
\description{
Addding indirect GO annotation
}
\details{
provided by a data.frame of GO TERM (column 1), GENE (column 2) and ONTOLOGY (optional) that
describes GO direct annotation,
this function will add indirect GO annotation of genes.
}
\author{
Yu Guangchuang
}
