% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMethods.R
\name{combineScores}
\alias{combineScores}
\title{combining similarity matrix to similarity score}
\usage{
combineScores(SimScores, combine)
}
\arguments{
\item{SimScores}{similarity matrix}

\item{combine}{combine method}
}
\value{
similarity value
}
\description{
Functions for combining similarity matrix to similarity score
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
