% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_organism}
\alias{get_organism}
\title{Get organism name from OrgDb object}
\usage{
get_organism(object)
}
\arguments{
\item{object}{OrgDb object or OrgDb package name}
}
\value{
Organism name
}
\description{
Get organism name from OrgDb object
}
\author{
Guangchuang Yu
}
