% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOSemSim-package.R
\docType{data}
\name{go_term_table}
\alias{go_term_table}
\alias{GO}
\alias{gotbl}
\title{Information content of GO terms}
\description{
These datasets are the information contents of GOterms.
}
\references{
Yu et al. (2010) GOSemSim: an R package for measuring semantic
similarity among GO terms and gene products \emph{Bioinformatics} (Oxford,
England), 26:7 976--978, April 2010. ISSN 1367-4803
\url{http://bioinformatics.oxfordjournals.org/cgi/content/abstract/26/7/976}
PMID: 20179076
}
\keyword{datasets}
