% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godata.R
\docType{class}
\name{GOSemSimDATA-class}
\alias{GOSemSimDATA-class}
\alias{show,GOSemSimDATA-method}
\title{Class "GOSemSimDATA"
This class stores IC and gene to go mapping for semantic similarity measurement}
\description{
Class "GOSemSimDATA"
This class stores IC and gene to go mapping for semantic similarity measurement
}
\section{Slots}{

\describe{
\item{\code{keys}}{gene ID}

\item{\code{ont}}{ontology}

\item{\code{IC}}{IC data}

\item{\code{geneAnno}}{gene to GO mapping}

\item{\code{metadata}}{metadata}
}}

\keyword{classes}
