\name{IC}
\docType{data}	
\alias{IC}
\alias{Info_Contents_MF_human}
\alias{Info_Contents_BP_human}
\alias{Info_Contents_CC_human}
\alias{Info_Contents_MF_rat}
\alias{Info_Contents_BP_rat}
\alias{Info_Contents_CC_rat}
\alias{Info_Contents_MF_mouse}
\alias{Info_Contents_BP_mouse}
\alias{Info_Contents_CC_mouse}
\alias{Info_Contents_MF_fly}
\alias{Info_Contents_BP_fly}
\alias{Info_Contents_CC_fly}
\alias{Info_Contents_MF_yeast}
\alias{Info_Contents_BP_yeast}
\alias{Info_Contents_CC_yeast}
\alias{Info_Contents_MF_zebrafish}
\alias{Info_Contents_BP_zebrafish}
\alias{Info_Contents_CC_zebrafish}
\alias{Info_Contents_MF_arabidopsis}
\alias{Info_Contents_BP_arabidopsis}
\alias{Info_Contents_CC_arabidopsis}
\alias{Info_Contents_MF_ecolik12}
\alias{Info_Contents_BP_ecolik12}
\alias{Info_Contents_CC_ecolik12}
\alias{Info_Contents_MF_worm}
\alias{Info_Contents_BP_worm}
\alias{Info_Contents_CC_worm}
\title{Information content of GO terms}
\description{  
	These datasets are the information contents of GOterms.
}	


\keyword{datasets}
