% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntDb.R
\name{load_onto}
\alias{load_onto}
\title{Load Ontology Database}
\usage{
load_onto(onto = "HDO")
}
\arguments{
\item{onto}{character. The ontology to load (e.g., "HDO").}
}
\value{
An \code{AnnotationDb} object.
}
\description{
Load Ontology Database
}
\keyword{internal}
