% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gene2GO.R
\name{loadGOMap}
\alias{loadGOMap}
\title{loadGOMap}
\usage{
loadGOMap(organism)
}
\arguments{
\item{organism}{one of supported organisms}
}
\value{
envir
}
\description{
loading GOMap to GOSemSimEnv
}
\author{
Yu Guangchuang
}

