% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadICdata}
\alias{loadICdata}
\title{Load IC data}
\usage{
loadICdata(organism, ont)
}
\arguments{
\item{organism}{"human"}

\item{ont}{"DO"}
}
\description{
Load Information Content data to DOSEEnv environment
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

