% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgeneSim.R
\name{mgeneSim}
\alias{mgeneSim}
\title{Pairwise semantic similarity for a list of genes}
\usage{
mgeneSim(
  genes,
  semData,
  measure = "Wang",
  drop = "IEA",
  combine = "BMA",
  verbose = TRUE,
  BPPARAM = NULL
)
}
\arguments{
\item{genes}{A list of Entrez gene IDs}

\item{semData}{GOSemSimDATA object}

\item{measure}{One of "Resnik", "Lin", "Rel", "Jiang", "TCSS" and "Wang" methods.}

\item{drop}{Evidence codes to drop; use \code{NULL} to keep all GO annotations}

\item{combine}{One of "max", "avg", "rcmax", "BMA" methods, used to combine multiple term scores.}

\item{verbose}{Whether to show a progress bar}

\item{BPPARAM}{optional \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object for
parallel pairwise similarity calculation. The default \code{NULL} uses the
original serial implementation.}
}
\value{
similarity matrix
}
\description{
Calculate pairwise semantic similarities for a given list of genes.
}
\details{
Parallel calculation is opt-in. With the default \code{BPPARAM = NULL},
\code{mgeneSim()} keeps the original serial behavior and can show a progress bar
when \code{verbose = TRUE}. When a \code{BPPARAM} object is supplied, pairwise
similarities are calculated through \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}} and the
progress bar is not shown.
}
\examples{
d <- godata('org.Hs.eg.db', ont = "MF", computeIC = FALSE)
mgeneSim(c("835", "5261", "241"), semData = d, measure = "Wang")
}
\seealso{
\code{\link[=goSim]{goSim()}} \code{\link[=mgoSim]{mgoSim()}} \code{\link[=geneSim]{geneSim()}} \code{\link[=mgeneSim]{mgeneSim()}} \code{\link[=clusterSim]{clusterSim()}} \code{\link[=mclusterSim]{mclusterSim()}}
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
