% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBlast2go.R
\name{read.blast2go}
\alias{read.blast2go}
\title{read.blast2go}
\usage{
read.blast2go(file)
}
\arguments{
\item{file}{BLAST2GO file}
}
\value{
a data frame with two columns: GO and Gene
}
\description{
given a BLAST2GO file, this function extracts the information from it and make it use for TERM2GENE.
}
