% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{load_OrgDb}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{yulab.utils}{\code{\link[yulab.utils]{load_OrgDb}}}
}}

