% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntDb.R
\name{set_auto_update}
\alias{set_auto_update}
\title{Control auto-update of ontology databases}
\usage{
set_auto_update(x = TRUE)
}
\arguments{
\item{x}{logical, whether to enable auto-update (default: TRUE)}
}
\description{
When enabled, outdated local ontology SQLite databases will be
automatically updated from remote mirrors.
}
