% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termSim.R
\name{termSim}
\alias{termSim}
\title{termSim}
\usage{
termSim(
  t1,
  t2,
  semData,
  method = c("Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS")
)
}
\arguments{
\item{t1}{Term vector}

\item{t2}{Term vector}

\item{semData}{GOSemSimDATA object}

\item{method}{One of "Wang", "Resnik", "Rel", "Jiang", "Lin", "TCSS"}
}
\value{
score matrix
}
\description{
Measure similarities between two term vectors.
}
\details{
Provide two term vectors, this function calculates their similarities.
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
