\name{wangMethod}
\alias{wangMethod}
\title{wangMethod}
\usage{
  wangMethod(ID1, ID2, ont = "BP")
}
\arguments{
  \item{ID1}{Ontology Term}

  \item{ID2}{Ontology Term}

  \item{ont}{Ontology}
}
\value{
  semantic similarity score
}
\description{
  Method Wang for semantic similarity measuring
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

