% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangMethod.R
\name{wangMethod_internal}
\alias{wangMethod_internal}
\title{wangMethod}
\usage{
wangMethod_internal(ID1, ID2, ont = "BP")
}
\arguments{
\item{ID1}{Ontology Term}

\item{ID2}{Ontology Term}

\item{ont}{Ontology}
}
\value{
semantic similarity score
}
\description{
Method Wang for semantic similarity measuring
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
