% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGOmap.R
\name{buildGOmap}
\alias{buildGOmap}
\title{buildGOmap}
\usage{
buildGOmap(x)
}
\arguments{
\item{x}{data.frame with two or three columns of GENE, GO and ONTOLOGY (optional)}
}
\value{
data.frame, GO annotation with direct and indirect annotation
}
\description{
Addding indirect GO annotation
}
\details{
provided by a data.frame of GENE (column 1), GO (column 2) and ONTOLOGY (optional) that
describes GO direct annotation, 
this function will add indirect GO annotation of genes.
}
\author{
Yu Guangchuang
}
