% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combineScores}
\alias{combineScores}
\title{combining similarity matrix to similarity score}
\usage{
combineScores(SimScores, combine)
}
\arguments{
\item{SimScores}{similarity matrix}

\item{combine}{combine method}
}
\value{
similarity value
}
\description{
Functions for combining similarity matrix to similarity score
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

