% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/godata.R
\name{godata}
\alias{godata}
\title{godata}
\usage{
godata(
  OrgDb = NULL,
  keytype = "ENTREZID",
  ont,
  computeIC = TRUE,
  processTCSS = FALSE,
  cutoff = NULL
)
}
\arguments{
\item{OrgDb}{OrgDb object}

\item{keytype}{keytype}

\item{ont}{one of 'BP', 'MF', 'CC'}

\item{computeIC}{logical, whether computer IC}

\item{processTCSS}{logical, whether to process TCSS}

\item{cutoff}{cutoff of TCSS}
}
\value{
GOSemSimDATA object
}
\description{
prepare GO DATA for measuring semantic similarity
}
\author{
Guangchuang Yu
}
