% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOSemSim-package.R
\docType{package}
\name{GOSemSim-package}
\alias{GOSemSim}
\alias{GOSemSim-package}
\title{GOSemSim: GO-terms Semantic Similarity Measures}
\description{
Semantic similarity analysis of Gene Ontology (GO) annotations provides a quantitative framework for comparing GO terms, gene products, and gene clusters. GOSemSim implements widely used information content- and graph-based similarity measures, including the methods of Resnik, Schlicker, Jiang, Lin, Wang and TCSS. It also provides utilities for preparing annotation data and combining term-level similarities into gene- and cluster-level scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/biomedical-knowledge-mining-book/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/GOSemSim/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com}
}

Other contributors:
\itemize{
  \item Alexey Stukalov \email{astukalov@gmail.com} [contributor]
  \item Pingfan Guo \email{1178431277@qq.com} [contributor]
  \item Chuanle Xiao \email{xiaochuanle@126.com} [contributor]
  \item Lluís Revilla Sancho \email{lluis.revilla@gmail.com} [contributor]
}

}
\keyword{internal}
