% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOSemSim-package.R
\docType{package}
\name{GOSemSim-package}
\alias{GOSemSim}
\alias{GOSemSim-package}
\title{GOSemSim: GO-terms Semantic Similarity Measures}
\description{
The semantic comparisons of Gene Ontology (GO) annotations provide quantitative ways to compute similarities between genes and gene groups, and have became important basis for many bioinformatics analysis approaches. GOSemSim is an R package for semantic similarity computation among GO terms, sets of GO terms, gene products and gene clusters. GOSemSim implemented five methods proposed by Resnik, Schlicker, Jiang, Lin and Wang respectively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/biomedical-knowledge-mining-book/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/GOSemSim/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

Other contributors:
\itemize{
  \item Alexey Stukalov \email{astukalov@gmail.com} [contributor]
  \item Pingfan Guo \email{1178431277@qq.com} [contributor]
  \item Chuanle Xiao \email{xiaochuanle@126.com} [contributor]
  \item Lluís Revilla Sancho \email{lluis.revilla@gmail.com} [contributor]
}

}
\keyword{internal}
