% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goSim.R
\name{goSim}
\alias{goSim}
\title{Semantic similarity between two GO terms}
\usage{
goSim(GOID1, GOID2, semData, measure = "Wang")
}
\arguments{
\item{GOID1}{GO ID 1}

\item{GOID2}{GO ID 2}

\item{semData}{GOSemSimDATA object}

\item{measure}{One of "Resnik", "Lin", "Rel", "Jiang", "TCSS" and "Wang" methods.}

\item{combine}{One of "max", "avg", "rcmax", "BMA" methods, used to combine multiple term scores.}
}
\value{
similarity
}
\description{
Given two GO IDs, calculate their semantic similarity.
}
\examples{
d <- godata('org.Hs.eg.db', ont = "MF", computeIC = FALSE)
goSim("GO:0004022", "GO:0005515", semData = d, measure = "Wang")
}
\references{
Yu et al. (2010) GOSemSim: an R package for measuring semantic similarity among GO terms and gene products.
Bioinformatics, 26(7): 976–978. https://academic.oup.com/bioinformatics/article/26/7/976/211914
}
\seealso{
\code{\link[=goSim]{goSim()}}, \code{\link[=mgoSim]{mgoSim()}}, \code{\link[=geneSim]{geneSim()}}, \code{\link[=mgeneSim]{mgeneSim()}}, \code{\link[=clusterSim]{clusterSim()}}, \code{\link[=mclusterSim]{mclusterSim()}}, \code{\link[=termSim]{termSim()}}

Other similarity: 
\code{\link{clusterSim}()},
\code{\link{geneSim}()},
\code{\link{mclusterSim}()},
\code{\link{mgeneSim}()},
\code{\link{mgoSim}()}
}
\concept{similarity}
