% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICMethods.R
\name{infoContentMethod}
\alias{infoContentMethod}
\title{information content based methods}
\usage{
infoContentMethod(ID1, ID2, method, godata)
}
\arguments{
\item{ID1}{Ontology Term}

\item{ID2}{Ontology Term}

\item{method}{one of "Resnik", "Jiang", "Lin" and "Rel".}

\item{godata}{GOSemSimDATA object}
}
\value{
semantic similarity score
}
\description{
Information Content Based Methods for semantic similarity measuring
}
\details{
implemented for methods proposed by Resnik, Jiang, Lin and Schlicker.
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
