% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termSim.R
\name{termSim}
\alias{termSim}
\title{termSim}
\usage{
termSim(t1, t2, semData, method = c("Wang", "Resnik", "Rel", "Jiang", "Lin"))
}
\arguments{
\item{t1}{term vector}

\item{t2}{term vector}

\item{semData}{GOSemSimDATA object}

\item{method}{one of "Wang", "Resnik", "Rel", "Jiang", and "Lin".}
}
\value{
score matrix
}
\description{
measuring similarities between two term vectors.
}
\details{
provide two term vectors, this function will calculate their similarities.
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
